refactor(app): extract pawwork workspace dialogs from layout.tsx
Slice 6 of the layout governance line (#1056). Pure extraction: moves the
DialogDeleteWorkspace / DialogResetWorkspace confirm-dialog components out of
pages/layout.tsx into pages/layout/pawwork-workspace-dialogs.tsx via a
createWorkspaceDialogs(input) factory. layout.tsx 1358 -> 1234.

No user-visible behavior change (DOM / copy / aria / i18n keys / styles
untouched). The two components are verbatim copies with an input. prefix; the
factory takes minimal Pick<> capability slices plus the slice-5 lifecycle
deleteWorkspace / resetWorkspace as input. The dialog.show(() => <DialogX .../>)
call sites in workspaceSidebarCtx are unchanged (the components are now
destructured from the factory).

Removed now-unused layout.tsx imports: Dialog, Button (only the extracted
dialogs used them).

These are presentational components; the logic-bearing deleteWorkspace /
resetWorkspace they call are already unit-tested in slice 5. Solid component
render tests in this repo require the browser-subprocess harness
(@/testing/browser-subprocess) because JSX templates need a client solid-js/web
build, which is disproportionate for two confirm dialogs — so this slice relies
on typecheck + the full app suite + codex equivalence review.

Verification: bun run typecheck 8/8; cd packages/app && bun test = 1667 pass;
eslint clean.

Refs #1056. #1053 (automation) also touches pages/layout.tsx and will rebase on top.
